Fixed PoissonSegment and added test

+++ src/segment_costs.jl
@@ -44,7 +44,7 @@ function ExponentialSegment(data::Array{Float64})
     return cost
 end
 
-function PoissonSegment(data::Array{Int64})
+function PoissonSegment(data::Array{Float64})
     cd = [0,cumsum( data )]
     cost(s::Int64, t::Int64) = -2*(cd[t+1]-cd[s+1]) * ( log(cd[t+1]-cd[s+1]) - log(t-s) - 1 )
     return cost
